Move the `DecodeStream` and `StreamsSequenceStream` from `src/core/stream.js` and into its own file
