+++ src/core/flate_stream.js
@@ -19,7 +19,7 @@
  * license.
  */
 
-import { DecodeStream } from "./stream.js";
+import { DecodeStream } from "./decode_stream.js";
 import { FormatError } from "../shared/util.js";
 
 // prettier-ignore
